UL-10 Reorganized app structure
Moved the ninja api instance into the views __init__ module.
Deleted account_management.api module.
